feat(github): add duplicate check
check jira issues for duplicates in github and ignore them

Refs: #8
