--- jihub.Github/Models/GitHubIssue.cs
@@ -1,14 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace jihub.Github.Models;
-
-public record GitHubIssue
-(
-    int Id,
-    int Number,
-    string Title,
-    string? Body,
-    GitHubMilestone Milestone,
-    GithubState State,
-    IEnumerable<GitHubLabel> Labels
-);
