+++ jihub.Github/Services/IGithubService.cs
@@ -4,7 +4,7 @@ namespace jihub.Github.Services;
 
 public interface IGithubService
 {
-    Task<GitHubInformation> GetMilestonesAndLabelsAsync(string owner, string repo, CancellationToken cts);
+    Task<GitHubInformation> GetRepoInformation(string owner, string repo, CancellationToken cts);
     Task<ICollection<GitHubLabel>> CreateLabelsAsync(string owner, string repo, IEnumerable<GitHubLabel> missingLabels, CancellationToken cts);
     Task<GitHubMilestone> CreateMilestoneAsync(string name, string owner, string repo, CancellationToken cts);
     Task CreateIssuesAsync(string owner, string repo, IEnumerable<CreateGitHubIssue> issues, CancellationToken cts);
